+++ NEWS.md
@@ -122,7 +122,7 @@ Updates for 2022 UNAIDS estimates:
 
 # first90 1.4.3
 
-* Patch: in function `add_ss_indices()`, add argument `type.convert(..., as.is = TRUE)` 
+* Patch: in function `add_ss_indices()`, add argument `utils::type.convert(..., as.is = TRUE)` 
   to suppress R 4.0 warning.
 * Bugfix: remove duplicate declaration of `double incrate_g[NG];` in `EPP_DIRECTINCID` incidence
   option. This did not affect any results because this option is not used in Shiny90 application;